Refactor ImageUploadView to socialhome app
Also make it use the correct form that creates an ImageUpload instance.

Refs: #120
