+++ socialhome/content/viewsets.py
@@ -1,11 +1,9 @@
-from rest_framework.generics import CreateAPIView
-from rest_framework.parsers import FormParser, MultiPartParser
-from rest_framework.permissions import BasePermission, SAFE_METHODS, IsAuthenticated
+from rest_framework.permissions import BasePermission, SAFE_METHODS
 from rest_framework.throttling import UserRateThrottle
 from rest_framework.viewsets import ModelViewSet
 
 from socialhome.content.models import Content
-from socialhome.content.serializers import ContentSerializer, ImageUploadSerializer
+from socialhome.content.serializers import ContentSerializer
 
 
 class IsOwnContentOrReadOnly(BasePermission):
@@ -42,14 +40,3 @@ class ContentViewSet(ModelViewSet):
 
     def perform_create(self, serializer):
         serializer.save(author=self.request.user.profile)
-
-
-class ImageUploadThrottle(UserRateThrottle):
-    scope = "image_upload"
-
-
-class ImageUploadView(CreateAPIView):
-    serializer_class = ImageUploadSerializer
-    permission_classes = (IsAuthenticated,)
-    parser_classes = (MultiPartParser, FormParser)
-    throttle_classes = (ImageUploadThrottle,)
